Factor out Types.Util.
drop unused re-export.
